Move SSA_stept definition to its own header file
This makes it easier to find it, and makes reading the
symex_target_equationt class and this one easier.
