+++ src/goto-checker/goto_symex_fault_localizer.h
@@ -44,7 +44,7 @@ protected:
   /// Collects the guards as \p localization_points up to \p failed_property_id
   /// and initializes fault_location_info, and returns the SSA step of
   /// the failed property
-  const symex_target_equationt::SSA_stept &collect_guards(
+  const SSA_stept &collect_guards(
     const irep_idt &failed_property_id,
     localization_pointst &localization_points,
     fault_location_infot &fault_location);
@@ -52,16 +52,15 @@ protected:
   // specify a localization point combination to check
   typedef std::vector<tvt> localization_points_valuet;
   bool check(
-    const symex_target_equationt::SSA_stept &failed_step,
+    const SSA_stept &failed_step,
     const localization_pointst &,
     const localization_points_valuet &);
 
   void update_scores(const localization_pointst &);
 
   // localization method: flip each point
-  void localize_linear(
-    const symex_target_equationt::SSA_stept &failed_step,
-    const localization_pointst &);
+  void
+  localize_linear(const SSA_stept &failed_step, const localization_pointst &);
 };
 
 #endif // CPROVER_GOTO_CHECKER_GOTO_SYMEX_FAULT_LOCALIZER_H
