+++ src/goto-symex/Makefile
@@ -15,6 +15,7 @@ SRC = auto_objects.cpp \
       show_program.cpp \
       show_vcc.cpp \
       slice.cpp \
+      ssa_step.cpp \
       symex_assign.cpp \
       symex_atomic_section.cpp \
       symex_builtin_functions.cpp \
