+++ src/goto-symex/equation_conversion_exceptions.h
@@ -23,7 +23,7 @@ class equation_conversion_exceptiont : public std::runtime_error
 public:
   equation_conversion_exceptiont(
     const std::string &message,
-    const symex_target_equationt::SSA_stept &step)
+    const SSA_stept &step)
     : runtime_error(message), step(step)
   {
     std::ostringstream error_msg;
@@ -40,7 +40,7 @@ public:
   }
 
 private:
-  symex_target_equationt::SSA_stept step;
+  SSA_stept step;
   std::string error_message;
 };
