+++ src/goto-symex/slice.cpp
@@ -51,7 +51,7 @@ void symex_slicet::slice(symex_target_equationt &equation)
     slice(*it);
 }
 
-void symex_slicet::slice(symex_target_equationt::SSA_stept &SSA_step)
+void symex_slicet::slice(SSA_stept &SSA_step)
 {
   get_symbols(SSA_step.guard);
 
@@ -109,8 +109,7 @@ void symex_slicet::slice(symex_target_equationt::SSA_stept &SSA_step)
   }
 }
 
-void symex_slicet::slice_assignment(
-  symex_target_equationt::SSA_stept &SSA_step)
+void symex_slicet::slice_assignment(SSA_stept &SSA_step)
 {
   PRECONDITION(SSA_step.ssa_lhs.id() == ID_symbol);
   const irep_idt &id=SSA_step.ssa_lhs.get_identifier();
@@ -124,8 +123,7 @@ void symex_slicet::slice_assignment(
     get_symbols(SSA_step.ssa_rhs);
 }
 
-void symex_slicet::slice_decl(
-  symex_target_equationt::SSA_stept &SSA_step)
+void symex_slicet::slice_decl(SSA_stept &SSA_step)
 {
   const irep_idt &id = to_symbol_expr(SSA_step.ssa_lhs).get_identifier();
 
@@ -152,7 +150,7 @@ void symex_slicet::collect_open_variables(
       it!=equation.SSA_steps.end();
       it++)
   {
-    const symex_target_equationt::SSA_stept &SSA_step=*it;
+    const SSA_stept &SSA_step = *it;
 
     get_symbols(SSA_step.guard);
